Broke out bundle version tracking to separate struct
