+++ relay/bundle_catalog.go
@@ -4,68 +4,8 @@ import (
 	"fmt"
 	"github.com/coreos/go-semver/semver"
 	"github.com/operable/go-relay/relay/config"
-	"sort"
 	"sync"
 )
-
-// List of semantic versions sorted in descending order
-type NewestFirst []*semver.Version
-
-func (nf NewestFirst) Len() int {
-	return len(nf)
-}
-
-func (nf NewestFirst) Swap(i, j int) {
-	nf[i], nf[j] = nf[j], nf[i]
-}
-
-func (nf NewestFirst) Less(i, j int) bool {
-	return nf[j].LessThan(*nf[i])
-}
-
-// VersionList is a list of semantic versions sorted in descending
-// order.
-type VersionList struct {
-	members NewestFirst
-}
-
-// NewVersionList constructs an empty version list
-func NewVersionList() *VersionList {
-	vl := &VersionList{
-		members: make(NewestFirst, 0),
-	}
-	return vl
-}
-
-// Add adds a new version to the list and ensures the list stays
-// sorted in descending order.
-func (vlp *VersionList) Add(version *semver.Version) {
-	vlp.members = append(vlp.members, version)
-	sort.Sort(vlp.members)
-}
-
-// Remove a version
-func (vlp *VersionList) Remove(version *semver.Version) {
-	candidate := version.String()
-	for i, v := range vlp.members {
-		if v.String() == candidate {
-			vlp.members = append(vlp.members[:i], vlp.members[i+1:]...)
-			break
-		}
-	}
-}
-
-func (vlp *VersionList) Len() int {
-	return len(vlp.members)
-}
-
-// Largest returns the largest version contained in the list
-func (vlp *VersionList) Largest() *semver.Version {
-	if vlp.members.Len() == 0 {
-		return nil
-	}
-	return vlp.members[0]
-}
 
 // BundleCatalog tracks installed and available bundles. Catalog
 // reads and writes are guarded by a single RWMutex.
@@ -106,9 +46,7 @@ func (bc *BundleCatalog) Add(bundle *config.Bundle) bool {
 			bc.versions[bundle.Name] = NewVersionList()
 		}
 		versions := bc.versions[bundle.Name]
-		fmt.Printf("1Before: %v\n", versions)
 		versions.Add(version)
-		fmt.Printf("1After: %v\n", versions)
 		bc.versions[bundle.Name] = versions
 		return true
 	}
@@ -153,7 +91,6 @@ func (bc *BundleCatalog) FindLatest(name string) *config.Bundle {
 	if latest == nil {
 		return nil
 	}
-	fmt.Printf("Versions: %v", versions)
 	key := bc.makeKey(name, latest.String())
 	return bc.bundles[key]
 }
